MDL-34328: Course dragdrop: Clone one resource drag handle instead of creating new ones
Instead of repeatedly calling get_drag_handle with the same parameters, and therefore creating a new (but identical) drag handle from scratch for each individual mod instance (resource/activity), just create one and perform a deep clone of it for each mod instance.  While this provides a small performance boost in all browsers, it's most noticeable in IE(8), since that's where everything is the slowest.