MDL-32880: moodle1 backup converter: add basic block handler
This handler processes the basic details of blocks - files and role assignments/overrides are currently ignored (though template inforef.xml and roles.xml files are created).  In order to use this generic handler, a block should extend this handler with a custom get_paths() function which attaches the "block" processor to the relevant path.