MDL-36204: rawurldecode file path before looking for it in the backup
Since the file paths returned by find_referenced_files are straight out of URLs, they need to be decoded before we can look for the actual files.  Otherwise, any files which include characters which require URL encoding (such as spaces) will not be migrated as they can't be found in the backup.