MDL-34491 backup: Large course won't backup due to POST limit
PHP's default max_input_vars=1000 causes (interactive) backup and restore
to fail (or be truncated), as it drops some of the submitted form fields.

This patch changes the schema step of the backup and restore UIs to use
arrays of fields in order to alleviate this issue (max_input_vars applies
to each level of an array, so this allows up to ~1K of each type of module
in any given course before hitting max_input_vars, which should suffice).