MDL-51918 gradereport: Fix regression caused by MDL-36606
Keyboard navigation (ctrl+up/down/left/right) with AJAX enabled ceased
functioning when editing mode is enabled after MDL-36606, as it now looks
for cells with a "clickable" class - but was only applying that class when
editing mode is NOT enabled.  This patch simply applies the "clickable"
class (where appropriate) irrespective of the editing mode's state.