MDL-50650 grade/import: Allow importing grades >100 from CSV
The CSV import process was aborting if any grade was over 100 (unless the
"unlimited grades" setting was enabled, in which case the "grade point
maximum" setting was used instead of 100). This does not make sense, as
grade updates are already bounded to the applicable grade item's min and
max. This patch removes the unnecessary checks, and displays a list of
adjusted grades (if any grades are adjusted to fit within the item's min
and max).